requirements.txt
numpy==1.19.5
pandas==1.3.0
matplotlib==3.4.3
scikit-learn==0.24.2
openpyxl==3.1.2
xlrd>=1.0.0
requests
unidecode==1.3.2
superset-api-client==0.6.0
sqlalchemy-1.4.39
psycopg2-2.9.3
jinja2